Make program module
